refactor: パッケージの移動
- Updated import paths across AI-related services and test files.
- Moved AI service files to a dedicated `ai` directory under `application/services`.
- Improved project organization and adherence to architectural clarity.
