+++ app/src/application/services/ai/AIService.test.ts
@@ -3,8 +3,11 @@
  */
 import { beforeEach, describe, expect, it } from 'vitest'
 
-import type { PuyoColor } from '../../domain/models/Puyo'
-import type { AIGameState, AISettings } from '../../domain/models/ai/types.ts'
+import type { PuyoColor } from '../../../domain/models/Puyo'
+import type {
+  AIGameState,
+  AISettings,
+} from '../../../domain/models/ai/types.ts'
 import { AIService } from './AIService'
 
 describe('AIService', () => {
