+++ app/src/application/services/ai/MoveGenerator.test.ts
@@ -3,8 +3,8 @@
  */
 import { describe, expect, it } from 'vitest'
 
-import type { PuyoColor } from '../../domain/models/Puyo'
-import type { AIGameState } from '../../domain/models/ai/types'
+import type { PuyoColor } from '../../../domain/models/Puyo'
+import type { AIGameState } from '../../../domain/models/ai/types'
 import { MoveGenerator } from './MoveGenerator'
 
 describe('MoveGenerator', () => {
