+++ app/src/infrastructure/di/DefaultContainer.ts
@@ -6,7 +6,7 @@ import type { StoragePort } from '../../application/ports/StoragePort'
 import type { TimerPort } from '../../application/ports/TimerPort'
 import { GameApplicationService } from '../../application/services/GameApplicationService'
 import { InputApplicationService } from '../../application/services/InputApplicationService'
-import { WorkerAIService } from '../../application/services/WorkerAIService.ts'
+import { WorkerAIService } from '../../application/services/ai/WorkerAIService.ts'
 import { ChainDetectionService } from '../../domain/services/ChainDetectionService'
 import { CollisionService } from '../../domain/services/CollisionService'
 import { PuyoSpawningService } from '../../domain/services/PuyoSpawningService'
